+++ platforms/swift/SnipsNlu/Dependencies/build.sh
@@ -4,7 +4,7 @@
 
 set -e
 
-VERSION="0.64.2"
+VERSION="0.64.3"
 SYSTEM=$(echo $1 | tr '[:upper:]' '[:lower:]')
 LIBRARY_NAME=libsnips_nlu_ffi
 LIBRARY_NAME_A=${LIBRARY_NAME}.a